Various fixes and improvements of the code
Faster chunk generation, loading.
Faster skylighting code.
Reworked a bit the chunk sending code, now the server sends a radius of 4 chunks at spawn.
PacketMap is a ConcurrentDictionary now, since it could be accessed from different threads (was causing the send of unhandled packets)
Deflaters are pooled for performance reasons, we shouldn't serialize with a lock the mapchunkpacket send.