Fix spawn chunk light not being calculated.
Now at the server start a spawn area of radius 4 is initialized and made Persistent so that it won't be never unloaded.
Now the chunk save also a bool that says if the skylight has to be recalculated, the old maps needs to be regenerated.
Remove some unnecessary debug log using the conditional compilation symbols.
Remove unused UpdateFirstChunks function since the chunks are pre-created/loaded on the world start.
Add a function to log on console or file on one line and with or without the header (date, loglevel).

Closes #131