Performance improvement on loading/saving chunk.
Implemented a versioning system so that future changes to saves won't need to have the map regenerated (if the code for loading is properly written).
Fixed a possible deadlock on save if the function is called simultaneously from two threads.
Note: change is not retroactive so the saves need to be deleted.